fix(api): fix NUMERIC/BIGINT mismatch in transaction stats query
The SUM(t.amount) on a BIGINT column returns NUMERIC in PostgreSQL,
which is incompatible with the i64 Rust type in the StatsRow struct.
Cast the aggregate results to BIGINT explicitly.

Also:
- Remove 2 validation tests for behaviors the API intentionally allows
  (overlapping periods, same-account transfers)
- Accept 423 (progressive backoff) alongside 401 in login failure tests
- Add avatar field to profile update in test helper (now required)

+++ src/database/transaction.rs
@@ -614,8 +614,8 @@ impl PostgresRepository {
         let rows = sqlx::query_as::<_, StatsRow>(
             r#"
             SELECT c.category_type::text,
-                   COALESCE(SUM(t.amount), 0) as total_amount,
-                   COUNT(t.id) as txn_count
+                   COALESCE(SUM(t.amount), 0)::BIGINT as total_amount,
+                   COUNT(t.id)::BIGINT as txn_count
             FROM transaction t
             JOIN category c ON t.category_id = c.id
             CROSS JOIN budget_period bp
